refactor(tests): consolidate desktop test-kernel suites onto their shared owners
Lane TC-desktop of the verified simplification survey (round 4).

- DesktopControlSystem.vitest.ts: delete 13 of 14 source-text-scraping
  tests, keeping only the embedded-browser URL/permission security
  policy assertions. The relaunch/quit pins keep owners at
  ElectronCompositionRoot.vitest.ts:162 and LifecycleHost.vitest.ts
  (runShutdown); the no-mirror invariant is behaviorally owned by
  DesktopWorkspaceIpc.vitest.ts ('accepts no editor dirty-state mirror
  from the renderer'), so the #9507 tombstones are doubly covered.
- DesktopThemeTokens.vitest.ts: drop the literal design-hex test
  (including its radial-gradient and info!==accent design-choice pins)
  and the two retired-token tombstone assertions.
- DesktopHostInteractions.vitest.ts: delete the suite that re-ran the
  shared createProgressHostInteractions factory (completing the round-2
  ruling PR #11441 half-executed). Five behaviors ported into
  ExtensionHostInteractions.vitest.ts: the await-tool-edits gate,
  sync-cancel-during-show, typed proposal overrides,
  dispose-with-stable-cause, and the reveal-before-external-inquiry
  registration (the fifth item #11441's verifier flagged as uncovered).
  Desktop wiring stays owned by DesktopAgentExecution.vitest.ts.
- DesktopDevScript.vitest.ts: delete the dev-launcher spawn
  choreography suite. The launch half is exercised on every real
  'npm run dev'; the replacement-args relaunch side of the supervisor
  contract (packages/desktop/src/main/index.ts, development supervisor
  comment) is deliberately given up, and the 2026-07-29 hermeticity
  investment in this test is knowingly discarded.
- BuildAliasConfig.vitest.ts: fold the two live positive pins
  (extension Vite shared-aliases import, esbuild tsconfig aliasing)
  into the alias-table equality test; delete the four tombstone
  assertions.
- DesktopSettingsIpc.vitest.ts: delete the two pure stub-delegation
  tests and the toggleTool/runInstallCommand halves of the Tools/LaTeX
  test; the LATEX_FORMATTER round trip survives. Routing rows arrive
  via typed spreads of controller handler dictionaries, and handler
  behavior keeps its own controller suites.
- DesktopUpdateChecker.vitest.ts: delete six shared-machine re-tests;
  port throttle, not-newer, and notify-throws cases into
  SemverUpdateCheck.vitest.ts so the shared suite is the single owner
  of runDailyUpdateCheck semantics. Keep the newer-release
  notify+persist case as the wiring smoke for
  DESKTOP_UPDATE_CHECK_LAST_NOTIFIED_VERSION, plus the URL-constant,
  isPackaged/env skips, and concurrency coalescing. Known residue: the
  undefined-version early return in updateCheck.ts stays unpinned
  (the refreshed:false gate covers the stamp).
