--- src/test-kernel/desktop/DesktopHostInteractions.vitest.ts
@@ -1,391 +1,0 @@
-import { describe, expect, it, vi } from 'vitest';
-
-import type {
-  BashSettlement,
-  PlanApprovalResult,
-  ProposalResult,
-  RetryResult,
-  UserQuestionSettlement,
-} from '@agent/runtime/HostInteractions';
-import type { SessionEvent } from '@agent/runtime/SessionEventHub';
-import {
-  createProgressHostInteractions,
-  type ProgressHostInteractionsOptions,
-} from '@controllers/progressView/backend/progressHostInteractions';
-import type { StreamTabId } from '@shared/schemas';
-import { SESSION_DISPOSED_CAUSE } from '@shared/copy/interactionCancellation';
-import { createTestSession } from '@test/support/sessionTestUtils';
-import type { ToolEditApprovalResult } from '@tools/approval/toolEditApproval';
-
-import { createRecordingApprovalHandlers } from '../progressView/approvalHandlerSetHarness';
-
-interface DesktopHostInteractions {
-  approvePendingDelegatedWork(
-    streamId: StreamTabId,
-    initiatingProposalId: string,
-  ): Promise<void>;
-  requestToolEditApproval(request: {
-    path: string;
-    originalContent: string;
-    proposedContent: string;
-    sourceTool: string;
-    streamId?: StreamTabId;
-  }): Promise<ToolEditApprovalResult>;
-  requestBashApproval(request: {
-    command: string;
-    streamId?: StreamTabId;
-  }): Promise<BashSettlement>;
-  requestPlanApproval(request: {
-    requestId: string;
-    streamId: StreamTabId;
-    goalEnabled: boolean;
-    plan: { objective: string };
-  }): Promise<unknown>;
-  requestAgentProposal(request: unknown): Promise<unknown>;
-  openExternalInquiry(request: {
-    requestId: string;
-    threadId: string;
-    streamId?: StreamTabId;
-  }): Promise<{ threadId: string }>;
-  submitBashDecision(requestId: string, decision: BashSettlement): boolean;
-  submitPlanDecision(requestId: string, decision: PlanApprovalResult): boolean;
-  submitProposalDecision(requestId: string, decision: ProposalResult): boolean;
-  submitUserQuestionDecision(
-    requestId: string,
-    decision: UserQuestionSettlement,
-  ): boolean;
-  cancel(selector?: {
-    streamId?: StreamTabId | null;
-    kind?: string;
-    cause?: string;
-  }): void;
-  dispose?(): void;
-}
-
-/** Reads the `requestId` passed to a handler's `.show()`, optionally for one stream. */
-function shownRequestId(
-  show: ReturnType<typeof vi.fn>,
-  streamId?: string,
-): string {
-  const call = streamId
-    ? show.mock.calls.find(([request]) => request.streamId === streamId)
-    : show.mock.calls[0];
-  const requestId = (call?.[0] as { requestId?: string } | undefined)
-    ?.requestId;
-  if (!requestId) throw new Error('Expected a captured requestId.');
-  return requestId;
-}
-
-/** Asserts the stream was registered without yanking the active tab (#8246). */
-function expectStreamRegistered(
-  sessionEvents: SessionEvent[],
-  streamId: StreamTabId,
-): void {
-  expect(sessionEvents).toContainEqual({
-    scope: 'session',
-    event: {
-      type: 'setActiveStream',
-      payload: {
-        streamId,
-        suppressViewSwitch: true,
-      },
-    },
-  });
-}
-
-async function createInteractions(
-  handlers = createRecordingApprovalHandlers(),
-) {
-  const presentationSink = { emit: vi.fn() };
-  const session = createTestSession();
-  const toolEditApprovals = {
-    approvePendingForStream: vi.fn(async (): Promise<void> => {}),
-    cancel: vi.fn(),
-    requestApproval: vi.fn(async () => ({
-      action: 'apply' as const,
-      appliedContent: 'new',
-    })),
-  };
-  const sessionEvents: SessionEvent[] = [];
-  session.events.subscribe((event) => sessionEvents.push(event), {
-    scope: 'session',
-  });
-
-  return {
-    interactions: createProgressHostInteractions({
-      interactions: presentationSink,
-      session,
-      getApprovalHandlers: () => handlers,
-      getToolEditApprovals: () =>
-        toolEditApprovals as unknown as ReturnType<
-          ProgressHostInteractionsOptions['getToolEditApprovals']
-        >,
-      setApprovalBypassState: vi.fn(),
-    }) as unknown as DesktopHostInteractions,
-    handlers,
-    presentationSink,
-    sessionEvents,
-    toolEditApprovals,
-  };
-}
-
-describe('createProgressHostInteractions on the desktop host', () => {
-  it('approves already-pending delegated work only in the selected stream', async () => {
-    const { interactions, handlers, toolEditApprovals } =
-      await createInteractions();
-    const current = interactions.requestAgentProposal({
-      requestId: 'proposal-current',
-      streamId: 'stream-a',
-    });
-    const parallel = interactions.requestAgentProposal({
-      requestId: 'proposal-parallel',
-      streamId: 'stream-a',
-    });
-    const bash = interactions.requestBashApproval({
-      command: 'lake build',
-      streamId: 'stream-a',
-    });
-    const other = interactions.requestBashApproval({
-      command: 'npm test',
-      streamId: 'stream-b',
-    });
-    let releaseToolEdits: (() => void) | undefined;
-    toolEditApprovals.approvePendingForStream.mockReturnValue(
-      new Promise<void>((resolve) => {
-        releaseToolEdits = resolve;
-      }),
-    );
-
-    let approvalCompleted = false;
-    const approval = interactions.approvePendingDelegatedWork(
-      'stream-a',
-      'proposal-current',
-    );
-    void approval.then(() => {
-      approvalCompleted = true;
-    });
-
-    await vi.waitFor(() =>
-      expect(toolEditApprovals.approvePendingForStream).toHaveBeenCalledWith(
-        'stream-a',
-      ),
-    );
-    expect(approvalCompleted).toBe(false);
-    releaseToolEdits?.();
-    await approval;
-
-    await expect(parallel).resolves.toEqual({ action: 'approve' });
-    await expect(bash).resolves.toEqual({ action: 'approve' });
-    expect(handlers.transport.proposal.dismiss).toHaveBeenCalledWith(
-      'proposal-parallel',
-    );
-
-    expect(
-      interactions.submitProposalDecision('proposal-current', {
-        action: 'approve',
-      }),
-    ).toBe(true);
-    const streamBRequestId = shownRequestId(
-      handlers.transport.bash.show,
-      'stream-b',
-    );
-    expect(
-      interactions.submitBashDecision(streamBRequestId, {
-        action: 'reject',
-      }),
-    ).toBe(true);
-    await expect(current).resolves.toEqual({ action: 'approve' });
-    await expect(other).resolves.toEqual({ action: 'reject' });
-  });
-
-  it('delegates tool edit approvals to the window controller', async () => {
-    const { interactions, toolEditApprovals } = await createInteractions();
-    const request = {
-      path: '/workspace/paper.tex',
-      originalContent: 'old',
-      proposedContent: 'new',
-      sourceTool: 'edit',
-      streamId: 'stream-a' as StreamTabId,
-    };
-
-    await expect(
-      interactions.requestToolEditApproval(request),
-    ).resolves.toEqual({ action: 'apply', appliedContent: 'new' });
-    // The controller owns its window session (options.session), so the call
-    // carries only the request.
-    expect(toolEditApprovals.requestApproval).toHaveBeenCalledWith(request);
-  });
-
-  it('rejects a plan decision for a pending bash request', async () => {
-    const handlers = createRecordingApprovalHandlers();
-    const { interactions, presentationSink, sessionEvents } =
-      await createInteractions(handlers);
-
-    const resultPromise = interactions.requestBashApproval({
-      command: 'echo hi',
-      streamId: 'stream-a' as StreamTabId,
-    });
-    const requestId = shownRequestId(handlers.transport.bash.show);
-
-    expect(
-      interactions.submitPlanDecision(requestId, { action: 'approve' }),
-    ).toBe(false);
-
-    expect(
-      interactions.submitBashDecision(requestId, { action: 'approve' }),
-    ).toBe(true);
-    await expect(resultPromise).resolves.toEqual({ action: 'approve' });
-    expect(presentationSink.emit).toHaveBeenCalledWith(
-      'requestEnsureProgressView',
-      {},
-    );
-    expect(presentationSink.emit).not.toHaveBeenCalledWith(
-      'setActiveStream',
-      expect.anything(),
-    );
-    expectStreamRegistered(sessionEvents, 'stream-a' as StreamTabId);
-  });
-
-  it('forwards a bash cancellation cause without user provenance', async () => {
-    const handlers = createRecordingApprovalHandlers();
-    const { interactions, toolEditApprovals } =
-      await createInteractions(handlers);
-
-    const resultPromise = interactions.requestBashApproval({
-      command: 'rm -rf build',
-      streamId: 'stream-a' as StreamTabId,
-    });
-
-    interactions.cancel({
-      streamId: 'stream-a' as StreamTabId,
-      cause: 'Stream resources released.',
-    });
-
-    await expect(resultPromise).resolves.toEqual({
-      action: 'reject',
-      cause: 'Stream resources released.',
-    });
-    expect(handlers.transport.bash.dismiss).toHaveBeenCalled();
-    expect(toolEditApprovals.cancel).toHaveBeenCalledWith({
-      streamId: 'stream-a',
-      cause: 'Stream resources released.',
-    });
-  });
-
-  it('can cancel synchronously while presenting a request', async () => {
-    const handlers = createRecordingApprovalHandlers();
-    const { interactions } = await createInteractions(handlers);
-    handlers.transport.bash.show.mockImplementation(() => {
-      interactions.cancel({
-        kind: 'bash',
-        streamId: 'stream-sync' as StreamTabId,
-        cause: 'Stopped during presentation.',
-      });
-    });
-
-    const result = interactions.requestBashApproval({
-      command: 'echo pending',
-      streamId: 'stream-sync' as StreamTabId,
-    });
-
-    await expect(result).resolves.toEqual({
-      action: 'reject',
-      cause: 'Stopped during presentation.',
-    });
-  });
-
-  it('routes tool-edit cancellation to the window controller', async () => {
-    const { interactions, toolEditApprovals } = await createInteractions();
-
-    interactions.cancel({
-      kind: 'toolEdit',
-      streamId: 'stream-a' as StreamTabId,
-      cause: 'Owning execution ended.',
-    });
-
-    expect(toolEditApprovals.cancel).toHaveBeenCalledWith({
-      kind: 'toolEdit',
-      streamId: 'stream-a',
-      cause: 'Owning execution ended.',
-    });
-  });
-
-  it('reveals the owning stream before showing an external inquiry', async () => {
-    const handlers = createRecordingApprovalHandlers();
-    const { interactions, presentationSink, sessionEvents } =
-      await createInteractions(handlers);
-
-    await expect(
-      interactions.openExternalInquiry({
-        requestId: 'inquiry-a',
-        threadId: 'thread-a',
-        streamId: 'stream-a' as StreamTabId,
-      }),
-    ).resolves.toEqual({ threadId: 'thread-a' });
-
-    expect(handlers.transport.externalInquiry.show).toHaveBeenCalledOnce();
-    expect(presentationSink.emit).toHaveBeenCalledWith(
-      'requestEnsureProgressView',
-      {},
-    );
-    expectStreamRegistered(sessionEvents, 'stream-a' as StreamTabId);
-  });
-
-  it('preserves typed proposal approval overrides', async () => {
-    const handlers = createRecordingApprovalHandlers();
-    const { interactions } = await createInteractions(handlers);
-
-    const resultPromise = interactions.requestAgentProposal({
-      requestId: 'proposal-a',
-      streamId: 'stream-a' as StreamTabId,
-      agentCategory: 'toolUse',
-      agent: 'demo-agent',
-    });
-
-    expect(
-      interactions.submitProposalDecision('proposal-a', {
-        action: 'approve',
-        model: 'openai:gpt-5',
-        agent: 'configured-agent',
-      }),
-    ).toBe(true);
-
-    await expect(resultPromise).resolves.toEqual({
-      action: 'approve',
-      model: 'openai:gpt-5',
-      agent: 'configured-agent',
-    });
-    expect(handlers.transport.proposal.dismiss).toHaveBeenCalledWith(
-      'proposal-a',
-    );
-  });
-
-  it('cancels all pending requests on dispose with a stable cause', async () => {
-    const handlers = createRecordingApprovalHandlers();
-    const { interactions } = await createInteractions(handlers);
-
-    const bashPromise = interactions.requestBashApproval({
-      command: 'echo hi',
-      streamId: 'stream-a' as StreamTabId,
-    });
-    const planPromise = interactions.requestPlanApproval({
-      requestId: 'plan-a',
-      streamId: 'stream-b' as StreamTabId,
-      goalEnabled: false,
-      plan: { objective: 'Prove the lemma.' },
-    });
-
-    interactions.dispose?.();
-
-    await expect(bashPromise).resolves.toEqual({
-      action: 'reject',
-      cause: SESSION_DISPOSED_CAUSE,
-    });
-    await expect(planPromise).resolves.toEqual({
-      action: 'reject',
-      cause: SESSION_DISPOSED_CAUSE,
-    });
-    expect(handlers.transport.bash.dismiss).toHaveBeenCalled();
-    expect(handlers.transport.planApproval.dismiss).toHaveBeenCalled();
-  });
-});
